Extracting WithExpectations module
