+++ app/models/exercise.rb
@@ -11,13 +11,11 @@ class Exercise < ActiveRecord::Base
   include WithSearch, WithTeaser,
           WithMarkup, WithAuthor,
           WithSubmissions, WithGuide,
-          WithLocale
+          WithLocale, WithExpectations
 
   acts_as_taggable
 
   belongs_to :language
-
-  serialize :expectations
 
   enum layout: [:editor_right, :editor_bottom, :no_editor, :scratchy]
 
@@ -62,18 +60,6 @@ class Exercise < ActiveRecord::Base
 
   def extra_code
     [guide.try(&:extra_code), self[:extra_code]].compact.join("\n")
-  end
-
-  def expectations_yaml
-    self.expectations.to_yaml
-  end
-
-  def expectations_yaml=(yaml)
-    self.expectations = YAML.load yaml
-  end
-
-  def expectations=(expectations)
-    self[:expectations] = expectations.map(&:stringify_keys)
   end
 
   private
