+++ spec/models/with_expectations_spec.rb
@@ -1,0 +1,34 @@
+describe WithExpectations do
+  let(:exercise) { create(:exercise) }
+
+  context 'when setting empty list' do
+    before { exercise.expectations = [] }
+
+    it { expect(exercise.expectations).to eq [] }
+  end
+
+  context 'when setting empty yaml' do
+    before { exercise.expectations_yaml = [].to_yaml }
+
+    it { expect(exercise.expectations).to eq [] }
+  end
+
+
+  context 'when setting non empty symbolized list' do
+    before { exercise.expectations = [{binding: 'foo', inspection: 'HasBinding'}] }
+
+    it { expect(exercise.expectations).to eq [{'binding' => 'foo', 'inspection' => 'HasBinding'}] }
+  end
+
+  context 'when setting non empty stringified list' do
+    before { exercise.expectations = [{'binding' => 'foo', 'inspection' => 'HasBinding'}] }
+
+    it { expect(exercise.expectations).to eq [{'binding' => 'foo', 'inspection' => 'HasBinding'}] }
+  end
+
+  context 'when setting non empty yaml' do
+    before { exercise.expectations_yaml = [{'binding' => 'foo', 'inspection' => 'HasBinding'}].to_yaml }
+
+    it { expect(exercise.expectations).to eq [{'binding' => 'foo', 'inspection' => 'HasBinding'}] }
+  end
+end
